+++ qap/climbnun.h
@@ -5,10 +5,10 @@
 #include <algorithm>   
 
 
-float nun_first(int ** D, int ** F, vector<int> * sol);
+long nun_first(int ** D, int ** F, vector<int> * sol);
 
-float nun_ME(int** D, int ** F, vector<int> * sol);
+long nun_ME(int** D, int ** F, vector<int> * sol);
 
-float nun_best(int** D, int ** F, vector<int> * sol);
+long nun_best(int** D, int ** F, vector<int> * sol);
 
 #endif
